update unit test cases
