+++ g/frame/gins/gins_database_test.go
@@ -24,8 +24,8 @@ test = "v=1"
         host     = "127.0.0.1"
         port     = "3306"
         user     = "root"
-        pass     = ""
-        # pass     = "12345678"
+        pass     = "12345678"
+        # pass     = ""
         name     = "test"
         type     = "mysql"
         role     = "master"
@@ -35,8 +35,8 @@ test = "v=1"
         host     = "127.0.0.1"
         port     = "3306"
         user     = "root"
-        pass     = ""
-        # pass     = "12345678"
+        pass     = "12345678"
+        # pass     = ""
         name     = "test"
         type     = "mysql"
         role     = "master"
