+++ g/frame/gins/gins_view_test.go
@@ -18,24 +18,24 @@ import (
 func Test_View(t *testing.T) {
     gtest.Case(t, func() {
         gtest.AssertNE(gins.View(), nil)
-        b, e := gins.View().ParseContent(`{{"1540822968" | date "Y-m-d H:i:s"}}`, nil)
+        b, e := gins.View().ParseContent(`{{"我是中国人" | substr 2 -1}}`, nil)
         gtest.Assert(e,         nil)
-        gtest.Assert(string(b), "2018-10-29 22:22:48")
+        gtest.Assert(string(b), "中国人")
     })
     gtest.Case(t, func() {
         tpl := "t.tpl"
-        err := gfile.PutContents(tpl, `{{"1540822968" | date "Y-m-d H:i:s"}}`)
+        err := gfile.PutContents(tpl, `{{"我是中国人" | substr 2 -1}}`)
         gtest.Assert(err, nil)
         defer gfile.Remove(tpl)
 
         b, e := gins.View().Parse("t.tpl", nil)
         gtest.Assert(e,         nil)
-        gtest.Assert(string(b), "2018-10-29 22:22:48")
+        gtest.Assert(string(b), "中国人")
     })
     gtest.Case(t, func() {
         path := fmt.Sprintf(`%s/%d`, gfile.TempDir(), gtime.Nanosecond())
         tpl  := fmt.Sprintf(`%s/%s`, path, "t.tpl")
-        err  := gfile.PutContents(tpl, `{{"1540822968" | date "Y-m-d H:i:s"}}`)
+        err  := gfile.PutContents(tpl, `{{"我是中国人" | substr 2 -1}}`)
         gtest.Assert(err, nil)
         defer gfile.Remove(tpl)
         err = gins.View().AddPath(path)
@@ -43,7 +43,7 @@ func Test_View(t *testing.T) {
 
         b, e := gins.View().Parse("t.tpl", nil)
         gtest.Assert(e,         nil)
-        gtest.Assert(string(b), "2018-10-29 22:22:48")
+        gtest.Assert(string(b), "中国人")
     })
 }
